splitup strds_renderer
